feat: `start` subcommand
