+++ src/arguments.rs
@@ -6,7 +6,7 @@ use tracing::level_filters::LevelFilter;
 
 /// The final word in Minecraft server scanners
 #[derive(Debug, Parser)]
-#[command(author, version = metadata::format(), about, long_about = None)]
+#[command(version = metadata::format(), about, subcommand_required = true)]
 pub struct Arguments {
     /// Amount of information to log
     #[arg(short, long, default_value = "info")]
@@ -22,6 +22,8 @@ pub struct Arguments {
 
 #[derive(Debug, Subcommand)]
 pub enum MainSubcommand {
+    Start,
+
     /// Manipulate openheimer configuration files
     Configuration {
         #[command(subcommand)]
